moved athletes to own app
