+++ Turnauswertung-py3/gymnastics/models/__init__.py
@@ -1,12 +1,10 @@
 from gymnastics.models.address import Address
-from gymnastics.models.athlete import Athlete
 from gymnastics.models.club import Club
 from gymnastics.models.discipline import Discipline
 from gymnastics.models.performance import Performance
 from gymnastics.models.squad import Squad
 from gymnastics.models.stream import Stream
-from gymnastics.models.team import Team
 from gymnastics.models.stream_discipline_join import StreamDisciplineJoin
-
+from gymnastics.models.team import Team
 from gymnastics.models.tournament import Tournament
-from gymnastics.models.athletes_import import AthletesImport
+
